Restructure package to run from root
