+++ install_prerequisites.py
@@ -7,7 +7,7 @@ import time
 
 def install_requirements():
     # Path to the requirements file inside the Wheatley project folder
-    req_file = os.path.join(os.path.dirname(__file__), "Wheatley", "python", "requirements.txt")
+    req_file = os.path.join(os.path.dirname(__file__), "wheatley", "requirements.txt")
     if not os.path.exists(req_file):
         print(f"Requirements file not found: {req_file}")
         sys.exit(1)
