+++ wheatley/__main__.py
@@ -1,0 +1,5 @@
+"""Entry point for running Wheatley as a module."""
+from .main import main
+
+if __name__ == "__main__":
+    main()
