+++ wheatley/llm/emotion_counter.json
