+++ wheatley/llm/llm_client.py
@@ -17,17 +17,17 @@ import tempfile
 
 #from local file google_agent import GoogleCalendarManager
 try:
-  from llm.google_agent import GoogleCalendarManager
+  from .google_agent import GoogleCalendarManager
 except ImportError:
   from google_agent import GoogleCalendarManager
 
 try:
-    from llm.spotify_agent import SpotifyAgent
+    from .spotify_agent import SpotifyAgent
 except ImportError:
     from spotify_agent import SpotifyAgent
 
-from llm.google_agent import GoogleAgent
-from llm.llm_client_utils import (
+from .google_agent import GoogleAgent
+from .llm_client_utils import (
     get_city_coordinates,
     get_quote,
     get_joke,
@@ -36,7 +36,7 @@ from llm.llm_client_utils import (
     build_tools,
     _load_config,
 )
-from utils.timing_logger import record_timing
+from ..utils.timing_logger import record_timing
 
 logging.basicConfig(level=logging.WARN)
 
@@ -480,18 +480,18 @@ class Functions:
 
     def write_long_term_memory(self, data: dict) -> str:
         """Persist ``data`` to the long term memory JSON file."""
-        from utils.long_term_memory import overwrite_memory
+        from ..utils.long_term_memory import overwrite_memory
         overwrite_memory(data, path=self.memory_path)
         return "memory written"
 
     def read_long_term_memory(self) -> dict:
         """Return the contents of the long term memory file."""
-        from utils.long_term_memory import read_memory
+        from ..utils.long_term_memory import read_memory
         return {"memory": read_memory(path=self.memory_path)}
 
     def edit_long_term_memory(self, index: int, data: dict) -> str:
         """Update the memory entry at ``index`` with ``data``."""
-        from utils.long_term_memory import edit_memory
+        from ..utils.long_term_memory import edit_memory
         success = edit_memory(index, data, path=self.memory_path)
         return "memory updated" if success else "memory index out of range"
