+++ wheatley/puppet.py
